+++ Src/FluentAssertions/AssertionExtensions.cs
@@ -921,6 +921,20 @@ public static class AssertionExtensions
 
 #endif
 
+#if NET6_0_OR_GREATER
+
+    /// <summary>
+    /// Returns a <see cref="TaskCompletionSourceAssertions"/> object that can be used to assert the
+    /// current <see cref="TaskCompletionSource"/>.
+    /// </summary>
+    [Pure]
+    public static TaskCompletionSourceAssertions Should(this TaskCompletionSource tcs)
+    {
+        return new TaskCompletionSourceAssertions(tcs);
+    }
+
+#endif
+
     /// <summary>
     /// Safely casts the specified object to the type specified through <typeparamref name="TTo"/>.
     /// </summary>
@@ -1038,7 +1052,7 @@ public static class AssertionExtensions
 
     /// <inheritdoc cref="Should(ExecutionTimeAssertions)" />
     [Obsolete("You are asserting the 'AndConstraint' itself. Remove the 'Should()' method directly following 'And'", error: true)]
-    public static void Should<TSubject>(this TaskCompletionSourceAssertions<TSubject> _)
+    public static void Should(this TaskCompletionSourceAssertionsBase _)
     {
         InvalidShouldCall();
     }